+++ proxy/src/main/java/com/velocitypowered/proxy/command/builtin/VelocityCommand.java
@@ -33,7 +33,7 @@ import com.velocitypowered.api.plugin.PluginDescription;
 import com.velocitypowered.api.proxy.ProxyServer;
 import com.velocitypowered.api.proxy.server.RegisteredServer;
 import com.velocitypowered.api.util.ProxyVersion;
-import com.velocitypowered.api.util.ServerBuildInfo;
+import com.velocitypowered.api.util.buildinfo.ServerBuildInfo;
 import com.velocitypowered.proxy.VelocityServer;
 import com.velocitypowered.proxy.util.InformationUtils;
 import java.io.BufferedWriter;
